+++ NEWS.md
@@ -1,10 +1,11 @@
 > Copyright © 2016 RTE Réseau de transport d’électricité
 
 # antaresRead 2.6.1 (devlopment)
-BUGFIXES : 
 
+BUGFIXES :  
+
+* `setSimulationPathAPI()` error message with bad id
 * `readInputThermal()` / `readInputRES()` fix "opts" parameter linked in sub function 
-
 
 # antaresRead 2.6.0